Added blood VFX and Enemy Damage
When the player hits the enemy, there is now a splash of blood that appears. The enemy has a visible health bar that rotates to face the player and the health bar updates whenever the enemy gets hit. When the enemy's health reaches 0, he ragdolls and the health bar disappears.